vk_bot/requirements.txt
vkbottle
requests
pillow
numpy